Add Retry on get_floating_ip fail
Base on the API documentation of OpenStack, the `addFloatingIp` action
upon the `/servers/{server-id}/action` endpoint returns 202-Accepted
on success. That means that the state of the database might not be
updated in some OpenStack implementations. In that sense, we have
added a retry interval in case the `/servers/{server-id}` endpoint
returns an empty list of floating ip addresses.